✏️ Rename: `sponsors.ts` to `SPONSORS.ts`
