+++ src/lib/components/sponsor_banner/SponsorBanner.svelte
@@ -1,6 +1,6 @@
 <!-- © 2022 REVATI -->
 <script lang="ts">
-	import { SPONSORS } from '$lib/data/sponsors';
+	import { SPONSORS } from '$lib/data/SPONSORS';
 	import Sponsor from './Sponsor.svelte';
 </script>
